refactor: moved files upload dialog content to a separate component
